requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
langgraph>=0.0.40
langchain-openai>=0.2.0
chromadb>=0.4.0
fastembed>=0.2.0
openai>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
pydantic-settings>=2.0.0
